Update README.md
## (B) Synthetic Learning Curve Extension
We can extend this to the "learning curve case" if we also require that in order to collect an unseen x, you will have to pay for it with the amount of fidelity that you would need to collect to arrive at the queried point.
This basically attaches a cost to the fidelity dimension, but not on switching configurations (the other dimension). Probably, this will result in the model searching for the ideal marginal which is a trivial problem --> so i removed it here.